+++ bin/src/cli/compile.rs
@@ -4,8 +4,9 @@ use kct_helper::io;
 use kct_kube::Filter;
 use kct_package::{Package, Release};
 use serde_json::Value;
+use std::env;
 use std::error::Error;
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 
 pub fn command() -> App<'static, 'static> {
 	SubCommand::with_name("compile")
@@ -17,10 +18,17 @@ pub fn command() -> App<'static, 'static> {
 				.required(true),
 		)
 		.arg(
-			Arg::with_name("values")
-				.short("f")
-				.long("values")
-				.help("Sets the values for the package")
+			Arg::with_name("input")
+				.short("i")
+				.long("input")
+				.help("Sets the input values for the package")
+				.takes_value(true),
+		)
+		.arg(
+			Arg::with_name("output")
+				.short("o")
+				.long("output")
+				.help("Where to save compiled yamls")
 				.takes_value(true),
 		)
 		.arg(
@@ -47,6 +55,9 @@ pub fn run(matches: &ArgMatches) -> Result<String, Box<dyn Error>> {
 	let values_from: Option<PathBuf> = matches.value_of("values").map(PathBuf::from);
 	let values = parse_values(&values_from)?;
 
+	let output: Option<PathBuf> = matches.value_of("output").map(PathBuf::from);
+	let output = ensure_output_exists(&output)?;
+
 	let package_from: PathBuf = matches.value_of("package").map(PathBuf::from).unwrap();
 	let package = Package::from_path(package_from)?;
 
@@ -62,12 +73,26 @@ pub fn run(matches: &ArgMatches) -> Result<String, Box<dyn Error>> {
 
 	let objects = kct_kube::find(&rendered, &filter)?;
 
-	let documents = objects
-		.iter()
-		.map(|object| serde_yaml::to_string(object).unwrap())
+	let documents: Vec<(PathBuf, String)> = objects
+		.into_iter()
+		.map(|(path, object)| (path, serde_yaml::to_string(&object).unwrap()))
 		.collect();
 
-	Ok(documents)
+	match output {
+		None => {
+			let stream: String = documents
+				.into_iter()
+				.map(|(_path, object)| object)
+				.collect();
+
+			Ok(stream)
+		}
+		Some(path) => {
+			write_objects(&path, documents)?;
+
+			Ok(format!("Objects written at \"{}\"", path.display()))
+		}
+	}
 }
 
 fn parse_values(path: &Option<PathBuf>) -> Result<Option<Value>, String> {
@@ -87,6 +112,38 @@ fn parse_values(path: &Option<PathBuf>) -> Result<Option<Value>, String> {
 			Ok(Some(parsed))
 		}
 	}
+}
+
+fn ensure_output_exists(path: &Option<PathBuf>) -> Result<Option<PathBuf>, String> {
+	match path {
+		None => Ok(None),
+		Some(path) => {
+			if path == &PathBuf::from("-") {
+				return Ok(None);
+			}
+
+			let base = env::current_dir().map_err(|err| err.to_string())?;
+
+			io::ensure_dir_exists(&base, path)
+				.map_err(|err| err.to_string())
+				.map(Some)
+		}
+	}
+}
+
+fn write_objects(root: &Path, objects: Vec<(PathBuf, String)>) -> Result<(), String> {
+	for (path, contents) in objects {
+		let target = {
+			let mut base = root.to_path_buf();
+
+			base.push(path.strip_prefix("/").unwrap());
+			base.with_extension("yaml")
+		};
+
+		io::write_contents(&target, &contents).map_err(|err| err.to_string())?;
+	}
+
+	Ok(())
 }
 
 fn as_paths(paths: &str) -> Vec<PathBuf> {