+++ crates/kct_kube/tests/find.rs
@@ -8,7 +8,7 @@ const MINIMAL_OBJECT: &str = r#"{
 	"apiVersion": "apps/v1"
 }"#;
 
-type Return = Result<Vec<Value>>;
+type Return = Result<Vec<(PathBuf, Value)>>;
 
 fn find_from(text: &str) -> Return {
 	let val: Value = serde_json::from_str(text).unwrap();
@@ -20,62 +20,157 @@ fn assert_invalid(err: Return) {
 	assert_eq!(err.unwrap_err(), Error::Invalid);
 }
 
-fn assert_valid(ok: Return, times: usize) {
+fn assert_objects(ok: Return, times: usize) {
 	assert!(ok.is_ok());
 
 	let object = serde_json::from_str(MINIMAL_OBJECT).unwrap();
 	let objects: Vec<Value> = iter::repeat(object).take(times).collect();
-	assert_eq!(ok.unwrap(), objects)
+
+	let rendered: Vec<Value> = ok
+		.unwrap()
+		.into_iter()
+		.map(|(_path, value)| value)
+		.collect();
+	assert_eq!(rendered, objects)
 }
 
-#[test]
-fn finds_objects() {
-	let json = MINIMAL_OBJECT;
-	let found = find_from(&json);
-	assert_valid(found, 1);
+fn assert_paths(ok: Return, paths: Vec<&str>) {
+	assert!(ok.is_ok());
 
-	let json = format!(r#"{{"a":{0}, "b":{0}}}"#, MINIMAL_OBJECT);
-	let found = find_from(&json);
-	assert_valid(found, 2);
-
-	let json = format!(
-		r#"{{"a":{{ "b": {0}, "c": {{ "d": {0}}}, "e": {0}}}}}"#,
-		MINIMAL_OBJECT
-	);
-	let found = find_from(&json);
-	assert_valid(found, 3);
+	let paths: Vec<PathBuf> = paths.into_iter().map(|s| PathBuf::from(s)).collect();
+	let rendered: Vec<PathBuf> = ok.unwrap().into_iter().map(|(path, _value)| path).collect();
+	assert_eq!(rendered, paths)
 }
 
-#[test]
-fn disallow_unclear_paths() {
-	let json = format!("[{0}, {0}]", MINIMAL_OBJECT);
-	let found = find_from(&json);
-	assert_invalid(found);
+mod objects {
+	use crate::{assert_invalid, assert_objects, find_from, MINIMAL_OBJECT};
 
-	let json = format!(r#"{{"a":{0}, "b":[{0}, {0}]}}"#, MINIMAL_OBJECT);
-	let found = find_from(&json);
-	assert_invalid(found);
+	#[test]
+	fn finds_objects() {
+		let json = MINIMAL_OBJECT;
+		let found = find_from(&json);
+		assert_objects(found, 1);
 
-	let json = format!(r#"{{"a":{{ "b": {0}, "c": [{0}, {0}]}}}}"#, MINIMAL_OBJECT);
-	let found = find_from(&json);
-	assert_invalid(found);
-}
+		let json = format!(r#"{{"a":{0}, "b":{0}}}"#, MINIMAL_OBJECT);
+		let found = find_from(&json);
+		assert_objects(found, 2);
 
-#[test]
-fn disallow_primitives() {
-	let values = [
-		"0",
-		"\"object\"",
-		"null",
-		&format!(r#"{{ "a":1,"b":{0} }}"#, MINIMAL_OBJECT),
-		&format!(
-			r#"{{ "a":{{ "b":{{ "c":null,"d":{0} }} }} }}"#,
+		let json = format!(
+			r#"{{"a":{{ "b": {0}, "c": {{ "d": {0}}}, "e": {0}}}}}"#,
 			MINIMAL_OBJECT
-		),
-		&format!(r#"{{ "a":{{ "b":{0} }}, "c": "str" }}"#, MINIMAL_OBJECT),
-	];
+		);
+		let found = find_from(&json);
+		assert_objects(found, 3);
+	}
 
-	for &json in values.iter() {
+	#[test]
+	fn disallow_primitives() {
+		let values = [
+			"0",
+			"\"object\"",
+			"null",
+			&format!(r#"{{ "a":1,"b":{0} }}"#, MINIMAL_OBJECT),
+			&format!(
+				r#"{{ "a":{{ "b":{{ "c":null,"d":{0} }} }} }}"#,
+				MINIMAL_OBJECT
+			),
+			&format!(r#"{{ "a":{{ "b":{0} }}, "c": "str" }}"#, MINIMAL_OBJECT),
+		];
+
+		for &json in values.iter() {
+			let found = find_from(&json);
+			assert_invalid(found);
+		}
+	}
+}
+
+mod paths {
+	use crate::{assert_invalid, assert_paths, find_from, MINIMAL_OBJECT};
+
+	#[test]
+	fn uses_prop_names_as_paths() {
+		let json = format!(
+			r#"{{"a":{{"b": {0}}}, "c": {{"d":{{"e":{0}}}}}, "b": {0}}}"#,
+			MINIMAL_OBJECT
+		);
+		let found = find_from(&json);
+
+		assert_paths(found, vec!["/a/b", "/b", "/c/d/e"])
+	}
+
+	#[test]
+	// https://github.com/CertainLach/jrsonnet/issues/52#issuecomment-878944237
+	fn orders_props_alphabetically() {
+		let json = format!(
+			r#"{{"z": {0}, "1": {0}, "01": {0}, "10": {0}, "2": {0}, "a": {0}}}"#,
+			MINIMAL_OBJECT
+		);
+		let found = find_from(&json);
+
+		assert_paths(found, vec!["/01", "/1", "/10", "/2", "/a", "/z"]);
+
+		let json = format!(
+			r#"{{"a": {{"c": {0}, "b": {0}}}, "c": {0}, "b": {{"z": {0}, "01": {0}, "a": {0}, "0": {0}}}}}"#,
+			MINIMAL_OBJECT
+		);
+		let found = find_from(&json);
+
+		assert_paths(
+			found,
+			vec!["/a/b", "/a/c", "/b/0", "/b/01", "/b/a", "/b/z", "/c"],
+		)
+	}
+
+	#[test]
+	fn allows_only_valid_and_clear_path_segments() {
+		let json = format!(r#"{{"01-object": {0}}}"#, MINIMAL_OBJECT);
+		let found = find_from(&json);
+		assert_paths(found, vec!["/01-object"]);
+
+		let json = format!(r#"{{"/": {0}}}"#, MINIMAL_OBJECT);
+		let found = find_from(&json);
+		assert_invalid(found);
+
+		let json = format!(r#"{{"a/b": {0}}}"#, MINIMAL_OBJECT);
+		let found = find_from(&json);
+		assert_invalid(found);
+
+		let json = format!(r#"{{".": {0}}}"#, MINIMAL_OBJECT);
+		let found = find_from(&json);
+		assert_invalid(found);
+
+		let json = format!(r#"{{"..": {0}}}"#, MINIMAL_OBJECT);
+		let found = find_from(&json);
+		assert_invalid(found);
+
+		let json = format!(r#"{{"some%thing": {0}}}"#, MINIMAL_OBJECT);
+		let found = find_from(&json);
+		assert_invalid(found);
+
+		let json = format!(r#"{{"this.complicates.filtering": {0}}}"#, MINIMAL_OBJECT);
+		let found = find_from(&json);
+		assert_invalid(found);
+
+		let json = format!(r#"{{"-start-alphanumeric": {0}}}"#, MINIMAL_OBJECT);
+		let found = find_from(&json);
+		assert_invalid(found);
+
+		let json = format!(r#"{{"end-alphanumeric-": {0}}}"#, MINIMAL_OBJECT);
+		let found = find_from(&json);
+		assert_invalid(found);
+	}
+
+	#[test]
+	fn disallow_unclear_paths() {
+		let json = format!("[{0}, {0}]", MINIMAL_OBJECT);
+		let found = find_from(&json);
+		assert_invalid(found);
+
+		let json = format!(r#"{{"a":{0}, "b":[{0}, {0}]}}"#, MINIMAL_OBJECT);
+		let found = find_from(&json);
+		assert_invalid(found);
+
+		let json = format!(r#"{{"a":{{ "b": {0}, "c": [{0}, {0}]}}}}"#, MINIMAL_OBJECT);
 		let found = find_from(&json);
 		assert_invalid(found);
 	}
@@ -121,7 +216,7 @@ mod filter {
 			};
 
 			let found = find_within_complex(&filter);
-			assert_valid(found, amount);
+			assert_objects(found, amount);
 		}
 
 		let found = find_within_minimal(&Filter {
@@ -129,7 +224,7 @@ mod filter {
 			only: vec![PathBuf::from("/")],
 		});
 
-		assert_valid(found, 1);
+		assert_objects(found, 1);
 	}
 
 	#[test]
@@ -154,7 +249,7 @@ mod filter {
 			};
 
 			let found = find_within_complex(&filter);
-			assert_valid(found, amount);
+			assert_objects(found, amount);
 		}
 
 		let found = find_within_minimal(&Filter {
@@ -162,7 +257,7 @@ mod filter {
 			only: vec![],
 		});
 
-		assert_valid(found, 0);
+		assert_objects(found, 0);
 	}
 
 	#[test]
@@ -194,7 +289,7 @@ mod filter {
 			let filter = Filter { only, except };
 
 			let found = find_within_complex(&filter);
-			assert_valid(found, amount);
+			assert_objects(found, amount);
 		}
 
 		let found = find_within_minimal(&Filter {
@@ -202,6 +297,6 @@ mod filter {
 			only: vec![PathBuf::from("/")],
 		});
 
-		assert_valid(found, 0);
+		assert_objects(found, 0);
 	}
 }