+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- documentation about motivation, usage, and package structure
+
+### Fixed
+
+- unwanted print on when validating values
+- wrong package version on `--version`
+
 ## [0.2.0] - 2020-10-23
 
 ### Added