+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.3.1] - 2021-07-03
 
 ### Added
 
@@ -20,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- unwanted print on when validating values
+- unwanted print when validating values
 - wrong package version on `--version`
 
 ## [0.2.0] - 2020-10-23
@@ -53,6 +53,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - package description from `kcp.json`
 - "extensibility" guarantees by forbidding non object paths to K8s objects
 
-[Unreleased]: https://github.com/bruno-delfino1995/kct/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/bruno-delfino1995/kct/compare/v0.3.1...HEAD
+[0.3.1]: https://github.com/bruno-delfino1995/kct/compare/v0.3.1...v0.2.0
 [0.2.0]: https://github.com/bruno-delfino1995/kct/compare/v0.2.0...v0.1.0
 [0.1.0]: https://github.com/bruno-delfino1995/kct/releases/tag/v0.1.0